Make NullPolicy<T> public
